uvarint: define the cut function
The new cut function is similar to the expect function.
The difference is it returns the remaining bytes
instead of the number of bytes read.

Callers will most likely want to decode a number of uvarints
and then process the rest of the bytes in a different way.
The cut function reduces the number of steps needed to accomplish this.

Tests for the new function have also been added.